refactor this code
