--- backend/fixMyArea/src/main/java/com/diviso/fixMyArea/repository/UserProfileRepository.java
@@ -1,15 +1,0 @@
-package com.diviso.fixMyArea.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.diviso.fixMyArea.model.UserProfile;
-
-public interface UserProfileRepository extends JpaRepository<UserProfile ,Long>{
-
-
-
-
-
-
-    
-}
